requirements.txt
google-genai>=1.0.0
notion-client==2.2.1
requests>=2.31.0
python-dotenv>=1.0.0
